requirements.txt
requests
beautifulsoup4
python-telegram-bot==21.3
schedule
python-dotenv

